nagaram/app/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
boto3==1.34.0